backend/requirements.txt
Django==3.2.16
psycopg2==2.9.9
python-dotenv==1.0.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
pillow==10.2.0
django-filter==24.1